functional w/ tiki sdk android lib imported
