+++ android/build.gradle
@@ -54,6 +54,9 @@ allprojects {
     repositories {
         google()
         mavenCentral()
+        maven {
+            url '/Users/maudi/Development/Repositories/TIKI/tiki-sdk-android/app/build/repo'
+        }
     }
 }
 
@@ -65,4 +68,5 @@ dependencies {
     testImplementation "junit:junit:$junitVersion"
     androidTestImplementation "androidx.test.ext:junit:$androidxJunitVersion"
     androidTestImplementation "androidx.test.espresso:espresso-core:$androidxEspressoCoreVersion"
+    implementation "com.mytiki:tiki-sdk-android:2.1.7"
 }
