+++ android/src/main/kotlin/com/mytiki/sdk/capacitor/Trail.kt
@@ -1,0 +1,5 @@
+package com.mytiki.sdk.capacitor
+
+class Trail {
+    val license: License = License()
+}
