+++ example/src/js/capacitor-welcome.js
@@ -1,6 +1,6 @@
 import { Camera } from '@capacitor/camera';
 import { SplashScreen } from '@capacitor/splash-screen';
-import { TikiSdk } from "@mytiki/tiki-sdk-capacitor";
+import { Tiki } from "@mytiki/tiki-sdk-capacitor";
 
 window.customElements.define(
   'capacitor-welcome',
@@ -93,7 +93,7 @@ window.customElements.define(
     connectedCallback() {
       const self = this;
       self.shadowRoot.querySelector('#take-photo').addEventListener('click', async function (e) {
-          const rsp = await TikiSdk.echo({value: "hm"})
+          const rsp = await Tiki.createLicense({value: "hm"})
           console.log(rsp.value)
 
           try {
