+++ src/index.ts
@@ -2,9 +2,9 @@ import { registerPlugin } from '@capacitor/core';
 
 import type { TikiSdkPlugin } from './definitions';
 
-const TikiSdk = registerPlugin<TikiSdkPlugin>('TikiSdk', {
+const Tiki = registerPlugin<TikiSdkPlugin>('TikiSdk', {
   web: () => import('./web').then(m => new m.TikiSdkWeb()),
 });
 
 export * from './definitions';
-export { TikiSdk };
+export { Tiki };
