+++ src/web.ts
@@ -3,8 +3,7 @@ import { WebPlugin } from '@capacitor/core';
 import type { TikiSdkPlugin } from './definitions';
 
 export class TikiSdkWeb extends WebPlugin implements TikiSdkPlugin {
-  async echo(options: { value: string }): Promise<{ value: string }> {
-    console.log('ECHO', options);
-    return options;
+  async createLicense(_: { value: string }): Promise<{ value: string }> {
+    throw this.unimplemented("Mobile Only.")
   }
 }
